+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Undo action upon the success add toast.
+
+### Fixed
+
+- Added message on **AddItemModal**
+- **pt** messages.
+
 ## [3.4.0] - 2020-08-20
 
 ### Added